requirements.txt
DateTime==4.3
feedparser==6.0.8
python-dotenv==0.19.2
requests==2.26.0
requests-file==1.5.1